requirements.txt
SpeechRecognition
pyaudio
google-generativeai
streamlit
python-dotenv
gTTS

-e .